DroneCAN: split out STM32 specific CAN code
ready for supporting other MCUs
