+++ Mcu/l431/Src/peripherals.c
@@ -12,9 +12,6 @@
 #include "ADC.h"
 #include "serial_telemetry.h"
 #include "targets.h"
-#if DRONECAN_SUPPORT
-#include <DroneCAN.h>
-#endif
 
 extern char bemf_timeout;
 
@@ -45,9 +42,6 @@ void initCorePeripherals(void)
 #endif
 #ifdef USE_INTERNAL_AMP
      init_OPAMP();
-#endif
-#if DRONECAN_SUPPORT
-     DroneCAN_Init();
 #endif
 }
